requirements.txt
Pillow
rembg
opencv-python
